refactor: rename file to drop-zone-input
