+++ components/common/index.ts
@@ -7,4 +7,4 @@ export { Switch } from "./Switch";
 export { default as Overlay } from "./Overlay";
 export { default as NavLink } from "./NavLink";
 export { default as ProgressBar } from "./ProgressBar";
-export * from "./DropZone";
+export * from "./DropZoneInput";
